Rename files for stub methods and test
